+++ NEWS.md
@@ -1,3 +1,12 @@
+# coffee 0.3.0
+* updated the vignette
+* added functions scissors and thinner for manipulation of MCMC output
+* enhanced strat code so it runs around 40% faster
+* previous strat runs can now be reloaded (using run=FALSE)
+* files are now read and written to faster (using fread, fwrite)
+* strat iterations can either be loaded into memory or written to files
+* added fit measure (and also offset for rings)
+
 # coffee 0.2.0
 * now links to the rintcal package (renamed from IntCal)
 * the strat function now handles undated levels, blocks of dates, and different types of gaps (of exactly known length, or according to a normal or gamma distribution)